fix(core/dfn): add contractDefaults, dfnPointers before running linters
